+++ examples/preact/src/Button/Button.story.tsx
@@ -32,6 +32,9 @@ function ButtonStory() {
         onEnter={() => {
           disabled = true;
         }}
+        onExit={() => {
+          disabled = false;
+        }}
       >
         <Button disabled={disabled}>{title}</Button>
       </Variant>