+++ packages/client/src/client/addons/events/EventsAddon.svelte
@@ -1,5 +1,6 @@
 <script>
   import CaretIcon from ':virtual/vitebook/icons/menu-caret?raw';
+  import ConsoleIcon from ':virtual/vitebook/icons/console?raw';
 
   import Addon from '../Addon.svelte';
   import { events } from './events';
@@ -13,16 +14,13 @@
   // Try to hide from the minifier.
   const l = window['console'];
 
-  let lastLogged;
+  function logEvent(event) {
+    l.log(event.ref);
+  }
+
   function openEvent(event) {
     const isOpen = open[event.id] ?? false;
-
     open = { ...open, [event.id]: !isOpen };
-
-    if (!isOpen && lastLogged !== event.ref) {
-      l.log(event.ref);
-      lastLogged = event.ref;
-    }
   }
 
   onDestroy(() => {
@@ -45,14 +43,25 @@
       <div class="addon__event">
         <button
           class="addon__event__detail"
+          aria-label="Show event details"
           on:pointerdown={() => openEvent(event)}
-          on:keydown={(e) => e.key === 'Enter' && openEvent(e)}
+          on:keydown={(e) => e.key === 'Enter' && openEvent(event)}
         >
           <span class="addon__event__caret" class:open={open[event.id]}>
             {@html CaretIcon}
           </span>
 
           <span class="addon__event__type">{event.ref.type}</span>
+
+          <button
+            class="addon__event__console"
+            aria-label="Log event to console"
+            on:pointerdown|stopPropagation={() => logEvent(event)}
+            on:keydown|stopPropagation={(e) =>
+              e.key === 'Enter' && logEvent(event)}
+          >
+            {@html ConsoleIcon}
+          </button>
         </button>
 
         <pre class:open={open[event.id]}>
@@ -83,10 +92,12 @@
   }
 
   .addon__event__detail {
+    position: relative;
     width: 100%;
     display: flex;
     align-items: center;
     padding: 0.375rem 0.5rem;
+    border-radius: 0.1rem;
   }
 
   .addon__event__type {
@@ -102,6 +113,17 @@
     display: inline-block;
     transform: rotate(-90deg);
     transition: transform 0.1s linear;
+  }
+
+  .addon__event__console {
+    position: absolute;
+    top: 0;
+    right: 0;
+    display: inline-block;
+    margin-left: auto;
+    z-index: 10;
+    height: 100%;
+    padding: 0.5rem;
   }
 
   .addon__event__caret.open {
@@ -132,6 +154,16 @@
     button:hover {
       color: var(--vbk--sidebar-item-hover-color);
       background-color: var(--vbk--sidebar-item-hover-bg-color);
+    }
+
+    .addon__event__console:hover {
+      color: var(--vbk--color-gray-200);
+      background-color: var(--vbk--color-gray-600);
+    }
+
+    :global(html.dark .addon__event__console:hover) {
+      color: var(--vbk--color-gray-600);
+      background-color: var(--vbk--color-gray-200);
     }
   }
 </style>