+++ packages/client/src/client/addons/events/events.ts
@@ -1,10 +1,13 @@
+import debounce from 'just-debounce-it';
+import throttle from 'just-throttle';
 import { writable } from 'svelte/store';
 
 import { stringifyEvent } from './stringifyEvent';
 
 export const events = writable<Events>([]);
 
-export function eventCallback(event: EventEntryRef) {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function eventCallback(event: any) {
   window.requestAnimationFrame(() => {
     events.update(($events) => {
       let stringifiedEvent;
@@ -20,6 +23,21 @@ export function eventCallback(event: EventEntryRef) {
       return $events.sort((a, b) => b.ref.timeStamp - a.ref.timeStamp);
     });
   });
+}
+
+export function throttledEventCallback(
+  interval = 0,
+  options?: {
+    leading?: boolean;
+    trailing?: boolean;
+  },
+) {
+  return throttle(eventCallback, interval, options);
+}
+
+export function debouncedEventCallback(wait = 0, callFirst?: boolean) {
+  // @ts-expect-error - .
+  return debounce(eventCallback, wait, callFirst);
 }
 
 export type EventEntry = {