+++ packages/create/bin/create-vitebook.js
@@ -111,10 +111,12 @@ async function main() {
 
   builder.pkg.addVitebookDependency('client');
   builder.pkg.addVitebookDependency('core');
+  builder.pkg.addDependency('svelte', '^3.43.1', { dev: true });
 
   switch (builder.framework) {
     case 'vue':
       builder.pkg.addDependency('vue', '^3.0.0');
+      builder.pkg.addDependency('@vue/compiler-sfc', '^3.2.26', { dev: true });
       builder.pkg.addVitebookDependency('vue');
       if (builder.hasFeature('markdown'))
         builder.pkg.addVitebookDependency('markdown-vue');
@@ -126,7 +128,6 @@ async function main() {
         builder.pkg.addVitebookDependency('markdown-preact');
       break;
     case 'svelte':
-      builder.pkg.addDependency('svelte', '^3.43.1', { dev: true });
       if (builder.hasFeature('markdown'))
         builder.pkg.addVitebookDependency('markdown-svelte');
   }