+++ packages/create/template-svelte/config.md.js
@@ -9,6 +9,8 @@ export default defineConfig(/** __GENERICS__ */)({
       appFile: 'App.svelte',
       svelte: {
         extensions: ['.svelte', '.md'],
+        // Remove if using `svelte-preprocess`.
+        useVitePreprocess: true,
       },
     }) /** __PLUGINS__ */,
   ],